I used the LineListener to trigger the SynthesisQueue Thread to pass the next speakable when the previous audio playing has finished.
Concerning the exception handling:
If an exception is ocurred in some method like the queueAudio() of MaryAudioFileOutput or the process of MaryClient,these exceptions are caught by the SynthesisQueue and are thrown as an Event to MarySynthesizedOutput.From there the event can be forwarded to ImplementationPlatform.
This way I can inform the voice browser about exceptions that may occur during the synthesis and audio playing processes.